Extract container_id logics out of the process class
