+++ src/ddprof_process.cc
@@ -12,62 +12,12 @@
 #include <unistd.h>
 #include <vector>
 
-#include <fstream>
-#include <optional>
-#include <regex>
-#include <string>
-
 namespace ddprof {
 constexpr auto k_max_buf_cgroup_link = 1024;
-
-constexpr static std::string_view UUID_SOURCE =
-    "[0-9a-f]{8}[-_][0-9a-f]{4}[-_][0-9a-f]{4}[-_][0-9a-f]{4}[-_][0-9a-f]{12}";
-constexpr static std::string_view CONTAINER_SOURCE = "[0-9a-f]{64}";
-constexpr static std::string_view TASK_SOURCE = "[0-9a-f]{32}-\\d+";
-
-namespace {
-std::optional<std::string> container_id_from_line(const std::string &line) {
-  std::smatch line_matches;
-  std::smatch container_matches;
-  const std::regex LINE_REGEX(R"(^\d+:[^:]*:(.+)$)");
-  const std::regex CONTAINER_REGEX(
-      "(" + std::string(UUID_SOURCE) + "|" + std::string(CONTAINER_SOURCE) +
-      "|" + std::string(TASK_SOURCE) + ")(?:.scope)? *$");
-
-  if (std::regex_search(line, line_matches, LINE_REGEX)) {
-    std::string match = line_matches[1].str();
-    if (std::regex_search(match, container_matches, CONTAINER_REGEX)) {
-      return container_matches[1].str();
-    }
-  }
-  return std::nullopt;
-}
-
-} // namespace
 
 std::string Process::format_cgroup_file(pid_t pid,
                                         std::string_view path_to_proc) {
   return string_format("%s/proc/%d/cgroup", path_to_proc.data(), pid);
-}
-
-DDRes extract_container_id(const std::string &filepath,
-                           ContainerId &container_id) {
-  container_id = std::nullopt;
-  std::ifstream cgroup_file(filepath);
-  if (!cgroup_file) {
-    // short lived pids can fail in this case
-    LG_DBG("Failed to open file: %s", filepath.data());
-    return ddres_warn(DD_WHAT_CGROUP);
-  }
-  std::string line;
-  while (std::getline(cgroup_file, line)) {
-    container_id = container_id_from_line(line);
-    if (container_id) {
-      return {};
-    }
-  }
-  container_id = "none";
-  return {};
 }
 
 const ContainerId &Process::get_container_id(std::string_view path_to_proc) {
@@ -75,10 +25,18 @@ const ContainerId &Process::get_container_id(std::string_view path_to_proc) {
     extract_container_id(format_cgroup_file(_pid, path_to_proc), _container_id);
     if (!_container_id) {
       // file can be gone (short lived pid ?)
+      // store a value to avoid further lookup
       _container_id = k_container_id_unknown;
     }
   }
   return _container_id;
+}
+
+Process::CGroupId_t Process::get_cgroup_ns(std::string_view path_to_proc) {
+  if (_cgroup_ns == kCGroupNsNull) {
+    read_cgroup_ns(_pid, path_to_proc, _cgroup_ns);
+  }
+  return _cgroup_ns;
 }
 
 DDRes Process::read_cgroup_ns(pid_t pid, std::string_view path_to_proc,
@@ -119,7 +77,7 @@ DDRes Process::read_cgroup_ns(pid_t pid, std::string_view path_to_proc,
 const ContainerId &ProcessHdr::get_container_id(pid_t pid, bool force) {
   // lookup cgroup
   static const ContainerId unknown_container_id =
-      ContainerId(Process::k_container_id_unknown);
+      ContainerId(k_container_id_unknown);
   auto it = _process_map.find(pid);
   if (it == _process_map.end()) {
     // new process, parse cgroup
@@ -131,6 +89,8 @@ const ContainerId &ProcessHdr::get_container_id(pid_t pid, bool force) {
       return unknown_container_id;
     }
   }
+  // uint64 is big enough that overflow is not a concern
+  // also consequence is just miss-labelling container-id
   ++(it->second._sample_counter);
   if (!force &&
       (it->second._sample_counter) < k_nb_samples_container_id_lookup) {
