+++ test/CMakeLists.txt
@@ -218,6 +218,7 @@ add_unit_test(
   ../src/build_id.cc
   ../src/common_mapinfo_lookup.cc
   ../src/common_symbol_lookup.cc
+  ../src/container_id.cc
   ../src/ddprof_process.cc
   ../src/ddprof_stats.cc
   ../src/dso.cc
@@ -254,6 +255,7 @@ add_unit_test(
   ../src/lib/allocation_tracker.cc
   ../src/base_frame_symbol_lookup.cc
   ../src/build_id.cc
+  ../src/container_id.cc
   ../src/common_mapinfo_lookup.cc
   ../src/common_symbol_lookup.cc
   ../src/ddprof_cmdline.cc
@@ -329,7 +331,8 @@ add_unit_test(tracepoint_config-ut tracepoint_config-ut.cc ../src/tracepoint_con
 
 add_unit_test(live_allocation-ut live_allocation-ut.cc ../src/live_allocation.cc)
 
-add_unit_test(ddprof_process-ut ddprof_process-ut.cc ../src/ddprof_process.cc)
+add_unit_test(ddprof_process-ut ddprof_process-ut.cc ../src/container_id.cc
+              ../src/ddprof_process.cc)
 
 add_benchmark(savecontext-bench savecontext-bench.cc ../src/lib/pthread_fixes.cc
               ../src/lib/savecontext.cc ../src/lib/saveregisters.cc)
